+++ examples/with-fusion-design/src/pages/Home/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Button, Calendar } from '@alifd/next';
+import { Button, Calendar, Icon } from '@alifd/next';
 import moment from 'moment';
 
 moment.locale('ru');
@@ -9,6 +9,7 @@ const Dashboard = () => {
   return (
     <>
       <h2>icejs with fusion design</h2>
+      <Icon type="alibaba" />
       <Button type="primary">Primary</Button>
       <Button type="secondary">secondary</Button>
       <Button type="normal">Normal</Button>