+++ packages/create-app-shared/src/weex/history.ts
@@ -1,10 +1,13 @@
 import { createMemoryHistory } from 'history';
 import createInitHistory from '../createInitHistory';
 import type { CreateHistory, InitHistory } from '../createInitHistory';
+import { setHistory } from '../storage';
 
 const createHistory: CreateHistory = () => {
   // Force memory history when env is weex or kraken
-  return createMemoryHistory();
+  const history = createMemoryHistory();
+  setHistory(history);
+  return history;
 };
 
 const initHistory: InitHistory = createInitHistory(createHistory);