+++ packages/plugin-stark-module/src/index.ts
@@ -7,10 +7,21 @@ import genRuntime from './genRuntime';
 import setExternals from './setExternals';
 import appendLifecycle from './appendLifecycle';
 
-const plugin: IPlugin = ({ onGetWebpackConfig, context, registerTask, onHook }, options) => {
+// TODO: remove this line next update
+// @ts-ignore
+const plugin: IPlugin = ({ onGetWebpackConfig, context, registerTask, onHook, registerUserConfig, hasRegistration }, options) => {
   const { command, userConfig, webpack, commandArgs } = context;
   const { minify: outerMinify, sourceMap: outerSourceMap } = (userConfig || {}) as IUserConfig;
 
+  const hasOutputDirRegistered = hasRegistration('outputDir', 'userConfig');
+  if (!hasOutputDirRegistered) {
+    registerUserConfig({
+      name: 'outputDir',
+      defaultValue: 'build',
+      validation: 'string',
+    });
+  }
+  
   const {
     moduleExternals,
     minify,