+++ packages/create-app-shared/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.0.2
+
+- [fix] set history when call createHistory in runtime plugins
+
+## 1.0.1
+
+- [fix] set history value after initHistory
+
 ## 1.0.0
 
 - [feat] refactor exports APIs